pyproject.toml
[tool.poetry]
name = "startup-deck-generator"
version = "0.1.0"
description = "AI-powered tool to generate startup pitch decks using CrewAI, Groq, and Streamlit."
authors = ["Your Name <you@example.com>"]
readme = "README.md"
packages = [{ include = "startup_deck_generator" }]

[tool.poetry.dependencies]
python = "^3.11"
crewai = "^0.22.8"
langchain-groq = "^0.1.2"
litellm = "^1.34.16"
streamlit = "^1.33.0"
markdown2 = "^2.4.13"
beautifulsoup4 = "^4.12.3"
reportlab = "^4.1.0"
python-dotenv = "^1.0.1"
pyyaml = "^6.0.1"

[tool.poetry.scripts]
startup-deck = "main:run_deck_generation"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
